requirements.txt
# ==========================================
# 배포용 최적화 의존성 패키지 (server5.py 실행용)
# ==========================================
# 
# 이 파일은 call_ex/server5.py를 실행하는데 필요한
# 최소한의 패키지만 포함합니다.
#
# 사용법:
#   pip install -r requirements2.txt
#
# ==========================================

# 웹 서버 & Twilio 통신
flask>=2.0.0
Flask-Sock>=0.6.0
twilio>=8.0.0
gevent>=21.0.0
gevent-websocket>=0.10.1
python-dotenv>=0.20.0

# 딥러닝 프레임워크
torch>=1.12.1
torchaudio>=2.0.2

# 오디오 처리
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.20.2,<2.0.0
scipy>=1.6.0

# ASR (Speech Recognition)
nemo_toolkit[asr]>=1.20.0
omegaconf>=2.1.0

# 키워드 추출 (Qwen3-1.7B)
transformers>=4.40.0

# Beam Search + Language Model (선택사항)
pyctcdecode>=0.5.0
kenlm
# kenlm 설치 실패 시: pip install https://github.com/kpu/kenlm/archive/master.zip

# ==========================================
# 설치 가이드:
# ==========================================
#
# 1. Python 환경 생성 (Python 3.8-3.10 권장)
#    conda create -n call_stt python=3.8
#    conda activate call_stt
#
# 2. 기본 패키지 설치
#    pip install -r requirements2.txt
#
# 3. GPU 사용 시 PyTorch 재설치 (CUDA 버전에 맞게)
#    pip uninstall -y torch torchvision torchaudio
#    
#    CUDA 12.1:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#    
#    CUDA 11.8:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# 4. kenlm 별도 설치 (Windows에서 실패할 경우)
#    pip install https://github.com/kpu/kenlm/archive/master.zip
#
# 5. 시스템 요구사항 (Ubuntu/Linux)
#    sudo apt-get update
#    sudo apt-get install -y libsndfile1 ffmpeg portaudio19-dev
#
# ==========================================
